Separate out Prometheus metric collector logic into sub-module
