+++ .github/CODEOWNERS
@@ -120,7 +120,8 @@
 /src/sinks/humio_logs.rs @jszwedko
 /src/sinks/splunk_hec.rs @ktff
 /src/sinks/kafka.rs @lukesteensen
-/src/sinks/prometheus.rs @lukesteensen
+/src/sinks/prometheus/collector.rs @bruceg
+/src/sinks/prometheus/sink.rs @lukesteensen
 /src/sinks/splunk_hec.rs @lukesteensen
 /src/sinks/statsd.rs @lukesteensen
 /src/sinks/tcp.rs @bruceg
