+++ src/sources/prometheus/mod.rs
@@ -175,7 +175,7 @@ mod test {
     use super::*;
     use crate::{
         config,
-        sinks::prometheus::PrometheusSinkConfig,
+        sinks::prometheus::sink::PrometheusSinkConfig,
         test_util::{next_addr, start_topology},
         Error,
     };
