+++ src/topology/mod.rs
@@ -721,7 +721,7 @@ mod tests {
 mod reload_tests {
     use crate::config::Config;
     use crate::sinks::console::{ConsoleSinkConfig, Encoding, Target};
-    use crate::sinks::prometheus::PrometheusSinkConfig;
+    use crate::sinks::prometheus::sink::PrometheusSinkConfig;
     use crate::sources::generator::GeneratorConfig;
     use crate::sources::splunk_hec::SplunkConfig;
     use crate::test_util::{next_addr, start_topology, wait_for_tcp};
